Assignment 4

+++ Lab Material/Lab 10– Web Pages using Next.js and React-Formss/Tutorial/lab10-bank-app/app/accounts/Accounts.jsx
@@ -13,6 +13,9 @@ export default function Accounts({ initialAccounts }) {
     }
     async function handleDeleteAccount(accountNo) {
         // 
+        const confirmed = confirm(`Are you sure you want to delete this account with no ${accountNo}?`)
+        if (!confirmed) return
+
         const url = `/api/accounts/${accountNo}`
         const response = await fetch(url, { method: 'DELETE' })
         const message = await response.json()
